split utils and refactoring
