+++ platform/ui/src/utils/index.js
@@ -1,0 +1,13 @@
+import capitalize from './capitalize';
+import colorManipulator from './colorManipulator';
+import getMockedStudies from './getMockedStudies';
+import getModalities from './getModalities';
+import getInstances from './getInstances';
+
+export {
+  capitalize,
+  colorManipulator,
+  getMockedStudies,
+  getModalities,
+  getInstances,
+};
